delete kurs

+++ src/main/java/gennadziy/KursyRepo.java
@@ -3,7 +3,9 @@ package gennadziy;
 import gennadziy.model.KursWalut;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import javax.persistence.criteria.CriteriaBuilder;
 
-public interface KursyRepo extends JpaRepository<KursWalut, Long> {
+
+public interface KursyRepo extends JpaRepository<KursWalut, Integer> {
 
 }
